requirements.txt
praw
OpenAI
python_dotenv